+++ dart/lib/scheme/scheme_fidelity.dart
@@ -14,17 +14,17 @@
 import 'dart:math' as math;
 
 import 'package:material_color_utilities/dislike/dislike_analyzer.dart';
+import 'package:material_color_utilities/dynamiccolor/dynamic_scheme.dart';
+import 'package:material_color_utilities/dynamiccolor/variant.dart';
 import 'package:material_color_utilities/hct/hct.dart';
 import 'package:material_color_utilities/palettes/tonal_palette.dart';
 import 'package:material_color_utilities/temperature/temperature_cache.dart';
-import 'dynamic_scheme.dart';
-import 'variant.dart';
 
 /// A scheme that places the source color in [Scheme.primaryContainer].
-/// 
+///
 /// Primary Container is the source color, adjusted for color relativity.
-/// It maintains constant appearance in light mode and dark mode. 
-/// This adds ~5 tone in light mode, and subtracts ~5 tone in dark mode. 
+/// It maintains constant appearance in light mode and dark mode.
+/// This adds ~5 tone in light mode, and subtracts ~5 tone in dark mode.
 ///
 /// Tertiary Container is the complement to the source color, using
 /// [TemperatureCache]. It also maintains constant appearance.