+++ dart/lib/scheme/scheme_rainbow.dart
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'package:material_color_utilities/dynamiccolor/dynamic_scheme.dart';
+import 'package:material_color_utilities/dynamiccolor/variant.dart';
 import 'package:material_color_utilities/hct/hct.dart';
 import 'package:material_color_utilities/palettes/tonal_palette.dart';
 import 'package:material_color_utilities/utils/math_utils.dart';
-import 'dynamic_scheme.dart';
-import 'variant.dart';
 
 /// A playful theme - the source color's hue does not appear in the theme.
 class SchemeRainbow extends DynamicScheme {